publisher/requirements.txt
faker==24.8.0
google-cloud-pubsub==2.23.0
